Fixed username and password swapped issue
Username and password are no longer swapped. The problem was in the constructor of User.java.

EMERGENCY TODO: We need to set the username to lowercase when we are checking to see if it exists in the database or not. We do NOT want to have usernames "Bob123" and "bob123" to both exist in the database because of this mistake.