Last commit for tonight. Got create course working. Edit course and Cancel course still remain.
Important notes on functionality.

The error trapping to make sure the same type of class isn't scheduled on the same day works. BUT, error trapping on the fields still needs to be implemented. We can either structure the date as just the day of hte week, mon or as a day/month/year. The later obviously complciates things a lot more and since the rubric makes no mention of full date and time we probably want to go with teh former.

Delete course is obviously pretty easy, you just generate all the classes of a given type and scan through by the date until a match is found.

Edit is slightly more interesting in that you have an intial date that you scan for and then a new date potentially. The same error trapping must be performed in the identical way as above such that you don't reschedule a class to an occupied date.