Major modifications to instructor UI (drop-down menu / Spinner.java)
Implemented a drop-down menu for the instructor to select the date for which they want to teach their course. Drop-down menus will significantly reduce errors and eliminate the need for error-trapping user input.

Set buttons and the scroll view side by side in a column format to improve UI readability.